parser: improve error reporting
In main.c, define the macro YY_CTX_LOCAL in order to have access to the
parsing context and the macro YY_INPUT in order to have access to the
current line number.  In order to make these macro visible to parser.c,
include parser.c in main.c.

Add the yyerror function for improved error report, showing the current
line and additional context.

Update the Makefile, removing the full.c file.