parser: improve diagnostic error context
Update the yyerror function to read from the input file, if the buffer
does not contain a complete line.

Add a caret (^) indicating where in the code the error was found.

Make the code printing the error context from the buffer more robust,
assuming that a LF ('\n') character may be present.

Remove extra HT ('\t') characters.

The implementation was adapted from
https://github.com/gpakosz/peg/blob/upstream/src/peg.c#L51, but using
Zig coding style and improving the readability of the C code.