Sync with langref grammar
Several changes were made directly to the grammar in the langref without
updating this copy, which is theoretically the canonical version.

This commit contains all changes made to the langref as of 0.15 which we
not yet reflected here. There have also been changes made here which are
not yet reflected in the langref which will be ported back soon. Manual
conflict resolution is fun :)